+++ app/src/main/java/com/kabouzeid/gramophone/appwidgets/AppWidgetSmall.java
@@ -1,7 +1,6 @@
 package com.kabouzeid.gramophone.appwidgets;
 
 import android.app.PendingIntent;
-import android.appwidget.AppWidgetManager;
 import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
@@ -24,9 +23,6 @@ import com.kabouzeid.gramophone.service.MusicService;
 import com.kabouzeid.gramophone.ui.activities.MainActivity;
 import com.kabouzeid.gramophone.util.Util;
 
-/**
- * @author Karim Abou Zeid (kabouzeid)
- */
 public class AppWidgetSmall extends BaseAppWidget {
     public static final String NAME = "app_widget_small";
 
@@ -41,24 +37,10 @@ public class AppWidgetSmall extends BaseAppWidget {
     }
 
     /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void onUpdate(final Context context, final AppWidgetManager appWidgetManager,
-                         final int[] appWidgetIds) {
-        defaultAppWidget(context, appWidgetIds);
-        final Intent updateIntent = new Intent(MusicService.APP_WIDGET_UPDATE);
-        updateIntent.putExtra(MusicService.EXTRA_APP_WIDGET_NAME, NAME);
-        updateIntent.putExtra(AppWidgetManager.EXTRA_APPWIDGET_IDS, appWidgetIds);
-        updateIntent.addFlags(Intent.FLAG_RECEIVER_REGISTERED_ONLY);
-        context.sendBroadcast(updateIntent);
-    }
-
-    /**
      * Initialize given widgets to default state, where we launch Music on
      * default click and hide actions if service not running.
      */
-    private void defaultAppWidget(final Context context, final int[] appWidgetIds) {
+    protected void defaultAppWidget(final Context context, final int[] appWidgetIds) {
         final RemoteViews appWidgetView = new RemoteViews(context.getPackageName(), R.layout.app_widget_small);
 
         appWidgetView.setViewVisibility(R.id.media_titles, View.INVISIBLE);
@@ -69,38 +51,6 @@ public class AppWidgetSmall extends BaseAppWidget {
 
         linkButtons(context, appWidgetView);
         pushUpdate(context, appWidgetIds, appWidgetView);
-    }
-
-    private void pushUpdate(final Context context, final int[] appWidgetIds, final RemoteViews views) {
-        final AppWidgetManager appWidgetManager = AppWidgetManager.getInstance(context);
-        if (appWidgetIds != null) {
-            appWidgetManager.updateAppWidget(appWidgetIds, views);
-        } else {
-            appWidgetManager.updateAppWidget(new ComponentName(context, getClass()), views);
-        }
-    }
-
-    /**
-     * Check against {@link AppWidgetManager} if there are any instances of this
-     * widget.
-     */
-    private boolean hasInstances(final Context context) {
-        final AppWidgetManager appWidgetManager = AppWidgetManager.getInstance(context);
-        final int[] mAppWidgetIds = appWidgetManager.getAppWidgetIds(new ComponentName(context,
-                getClass()));
-        return mAppWidgetIds.length > 0;
-    }
-
-    /**
-     * Handle a change notification coming over from
-     * {@link MusicService}
-     */
-    public void notifyChange(final MusicService service, final String what) {
-        if (hasInstances(service)) {
-            if (MusicService.META_CHANGED.equals(what) || MusicService.PLAY_STATE_CHANGED.equals(what)) {
-                performUpdate(service, null);
-            }
-        }
     }
 
     /**
